fix: remove main router and move whoToFollow function to main filr (but no router)
