+++ src/server/api/main.ts
@@ -1,0 +1,35 @@
+import { clerkClient } from "@clerk/nextjs/server"
+import { getUserFollowList } from "./follow"
+import { ProfileExtend } from "~/components/profilePage/types"
+import { getFullName } from "~/utils/helpers"
+
+const MAS_USERS_TO_FOLLOW = 10
+
+export const whoToFollow = async (userId: string | null) => {
+	const usersToFollow = await clerkClient.users.getUserList({
+		orderBy: "-created_at",
+		limit: MAS_USERS_TO_FOLLOW,
+	})
+
+	let users = [...usersToFollow]
+	if (userId) {
+		const signInUser = await clerkClient.users.getUser(userId)
+		const follows = await getUserFollowList(signInUser.id)
+		if (follows) {
+			users = users.filter((user) => !follows.includes(user.id))
+		}
+	}
+
+	return users.map((user) => {
+		const extended = user.publicMetadata.extended as ProfileExtend | null
+
+		const fullName = getFullName(user.firstName, user.lastName)
+		return {
+			id: user.id,
+			username: user.username,
+			fullName,
+			profileImageUrl: user.profileImageUrl,
+			bio: extended?.bio ?? null,
+		}
+	})
+}
