+++ src/server/api/root.ts
@@ -2,7 +2,6 @@ import { createTRPCRouter } from "~/server/api/trpc"
 import { postsRouter } from "~/server/api/routers/posts"
 import { profileRouter } from "./routers/profile"
 import { followRouter } from "./routers/follow"
-import { mainRouter } from "./routers/main"
 
 /**
  * This is the primary router for your server.
@@ -13,7 +12,6 @@ export const appRouter = createTRPCRouter({
 	posts: postsRouter,
 	profile: profileRouter,
 	follow: followRouter,
-	main: mainRouter,
 })
 
 // export type definition of API
